ImageAnalyzerService/requirements.txt
fastapi
uvicorn
pillow
google-cloud-vision
python-dotenv
python-multipart